+++ lib/xray_native.dart
@@ -1,0 +1,1 @@
+export 'src/native/xray_core_flutter_native.dart';